requirements.txt
flask
pyserial
google-genai